equipment-visualizer/backend/requirements.txt
Django==6.0.2
djangorestframework==3.15.2
django-cors-headers==4.6.0
pandas==2.2.2
